requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
git-filter-repo==2.47.0
google-auth==2.40.0
google-auth-oauthlib==1.2.2
greenlet==3.2.1
gspread==6.2.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
response==0.5.0
rsa==4.9.1
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
